requirements.txt
groq
sounddevice
numpy
scipy
keyboard
pyautogui
pyperclip
python-dotenv
